update search for nonlocalized string datatype and extract
