+++ arches/app/datatypes/core/__init__.py
@@ -1,1 +1,2 @@
 from .geojson_feature_collection import *
+from .non_localized_string import *
