+++ arches/app/datatypes/datatypes.py
@@ -2,6 +2,7 @@ import copy
 import uuid
 import json
 import decimal
+from arches.app.datatypes.core.util import get_value_from_jsonld
 from arches.app.utils.file_validator import FileValidator
 import filetype
 import base64
@@ -453,104 +454,6 @@ class StringDataType(BaseDataType):
         tile_language_codes = set(tile.data[nodeid].keys())
         for code in all_language_codes - tile_language_codes:
             tile.data[nodeid][code] = {"value": "", "direction": direction_lookup[code]}
-
-
-class NonLocalizedStringDataType(BaseDataType):
-    def validate(
-        self,
-        value,
-        row_number=None,
-        source=None,
-        node=None,
-        nodeid=None,
-        strict=False,
-        **kwargs,
-    ):
-        errors = []
-        try:
-            if value is not None:
-                value.upper()
-        except:
-            message = _("This is not a string")
-            error_message = self.create_error_message(
-                value, source, row_number, message
-            )
-            errors.append(error_message)
-        return errors
-
-    def clean(self, tile, nodeid):
-        if tile.data[nodeid] in ["", "''"]:
-            tile.data[nodeid] = None
-
-    def append_to_document(self, document, nodevalue, nodeid, tile, provisional=False):
-        if nodevalue is not None:
-            val = {
-                "string": nodevalue,
-                "language": "",
-                "nodegroup_id": tile.nodegroup_id,
-                "provisional": provisional,
-            }
-            document["strings"].append(val)
-
-    def transform_export_values(self, value, *args, **kwargs):
-        if value is not None:
-            return value
-
-    def get_search_terms(self, nodevalue, nodeid=None):
-        terms = []
-
-        if nodevalue is not None:
-            if settings.WORDS_PER_SEARCH_TERM is None or (
-                len(nodevalue.split(" ")) < settings.WORDS_PER_SEARCH_TERM
-            ):
-                terms.append(SearchTerm(value=nodevalue, lang=""))
-        return terms
-
-    def append_search_filters(self, value, node, query, request):
-        try:
-            if value["op"] == "null" or value["op"] == "not_null":
-                self.append_null_search_filters(value, node, query, request)
-            elif value["val"] != "":
-                match_type = "phrase_prefix" if "~" in value["op"] else "phrase"
-                match_query = Match(
-                    field="tiles.data.%s" % (str(node.pk)),
-                    query=value["val"],
-                    type=match_type,
-                )
-                if "!" in value["op"]:
-                    query.must_not(match_query)
-                    query.filter(Exists(field="tiles.data.%s" % (str(node.pk))))
-                else:
-                    query.must(match_query)
-        except KeyError as e:
-            pass
-
-    def is_a_literal_in_rdf(self):
-        return True
-
-    def to_rdf(self, edge_info, edge):
-        # returns an in-memory graph object, containing the domain resource, its
-        # type and the string as a string literal
-        g = Graph()
-        if edge_info["range_tile_data"] is not None:
-            g.add((edge_info["d_uri"], RDF.type, URIRef(edge.domainnode.ontologyclass)))
-            g.add(
-                (
-                    edge_info["d_uri"],
-                    URIRef(edge.ontologyproperty),
-                    Literal(edge_info["range_tile_data"]),
-                )
-            )
-        return g
-
-    def from_rdf(self, json_ld_node):
-        # returns the string value only
-        # FIXME: Language?
-        value = get_value_from_jsonld(json_ld_node)
-        try:
-            return value[0]
-        except (AttributeError, KeyError) as e:
-            pass
 
 
 class NumberDataType(BaseDataType):
@@ -2522,21 +2425,3 @@ class AnnotationDataType(BaseDataType):
             }
         }
         return mapping
-
-
-def get_value_from_jsonld(json_ld_node):
-    try:
-        language = json_ld_node[0].get("@language")
-        if language is None:
-            language = get_language()
-        return (json_ld_node[0].get("@value"), language)
-    except KeyError as e:
-        try:
-            language = json_ld_node.get("@language")
-            if language is None:
-                language = get_language()
-            return (json_ld_node.get("@value"), language)
-        except AttributeError as e:
-            return
-    except IndexError as e:
-        return
